interactive bots: Extracted GameAdapter Class.
