+++ zulip_bots/zulip_bots/bots/connect_four/controller.py
@@ -17,7 +17,10 @@ class ConnectFourModel(object):
         [0, 0, 0, 0, 0, 0, 0]]
 
     current_board = blank_board
-
+    
+    def parse_move(self, move):
+        return int(move) - 1
+    
     def update_board(self, board):
         self.current_board = deepcopy(board)
